move inDelta functions to util and remove dependency on errors
