+++ math/dec.go
@@ -44,6 +44,7 @@ var (
 	ErrOverflow           = errorsmod.Register(mathCodespace, 5, "overflow")
 	ErrNaN                = errorsmod.Register(mathCodespace, 6, "NaN not permitted in this context")
 	ErrNotMatchingLength  = errorsmod.Register(mathCodespace, 7, "slices are not of the same length")
+	ErrOutOfRange         = errorsmod.Register(mathCodespace, 8, "value is out of range")
 )
 
 // The number 0 encoded as Dec
@@ -739,62 +740,4 @@ func (x Dec) Reduce() (Dec, int) {
 	y := Dec{}
 	_, n := y.dec.Reduce(&x.dec)
 	return y, n
-}
-
-// helper function for test suites that want to check
-// if some math is within a delta
-func InDelta(expected, result Dec, epsilon Dec) (bool, error) {
-	if expected.IsNaN() || result.IsNaN() {
-		return false, errorsmod.Wrap(ErrNaN, "cannot compare NaN")
-	}
-	delta, err := expected.Sub(result)
-	if err != nil {
-		return false, nil
-	}
-	deltaAbs, err := delta.Abs()
-	if err != nil {
-		return false, errorsmod.Wrap(err, "error getting absolute value")
-	}
-	compare := deltaAbs.Cmp(epsilon)
-	if compare == LessThan || compare == EqualTo {
-		return true, nil
-	}
-	return false, nil
-}
-
-// Helper function to compare two slices of alloraMath.Dec within a delta
-func SlicesInDelta(a, b []Dec, epsilon Dec) (bool, error) {
-	lenA := len(a)
-	if lenA != len(b) {
-		return false, errorsmod.Wrap(ErrNotMatchingLength, "cannot check if slices are within delta")
-	}
-	for i := 0; i < lenA; i++ {
-		// for performance reasons we do not call InDelta
-		// pass by copy causes this to run slow af for large slices
-		delta, err := a[i].Sub(b[i])
-		if err != nil {
-			return false, errorsmod.Wrapf(err, "error subtracting %v from %v", b[i], a[i])
-		}
-		deltaAbs, err := delta.Abs()
-		if err != nil {
-			return false, errorsmod.Wrap(err, "error getting absolute value")
-		}
-		if deltaAbs.Cmp(epsilon) == GreaterThan {
-			return false, nil
-		}
-	}
-	return true, nil
-}
-
-// Generic Sum function, given an array of values returns its sum
-func SumDecSlice(x []Dec) (Dec, error) {
-	sum := ZeroDec()
-	var err error
-	for _, v := range x {
-		sum, err = sum.Add(v)
-		if err != nil {
-			return Dec{}, errorsmod.Wrapf(err, "error adding %v + %v", v, sum)
-		}
-	}
-	return sum, nil
 }
